+++ mypy/test/data.py
@@ -6,6 +6,7 @@ import re
 from os import remove, rmdir
 import shutil
 
+import pytest
 from typing import Callable, List, Tuple, Set, Optional
 
 from mypy.myunit import TestCase, SkipTestCaseException
@@ -218,15 +219,6 @@ class TestItem:
         self.line = line
 
 
-class DataSuite:
-    @classmethod
-    def cases(cls) -> List[DataDrivenTestCase]:
-        return []
-
-    def run_case(self, testcase: DataDrivenTestCase) -> None:
-        raise NotImplementedError
-
-
 def parse_test_data(l: List[str], fnam: str) -> List[TestItem]:
     """Parse a list of lines that represent a sequence of test items."""
 
@@ -345,3 +337,79 @@ def fix_win_path(line: str) -> str:
         filename, lineno, message = m.groups()
         return '{}:{}{}'.format(filename.replace('/', '\\'),
                                 lineno or '', message)
+
+
+##
+#
+# pytest setup
+#
+##
+
+
+def pytest_addoption(parser):
+    group = parser.getgroup('mypy')
+    group.addoption('--update-data', action='store_true', default=False,
+                    help='Update test data to reflect actual output'
+                         ' (supported only for certain tests)')
+
+
+def pytest_pycollect_makeitem(collector, name, obj):
+    if not isinstance(obj, type) or not issubclass(obj, DataSuite):
+        return None
+    #os.write(3, ('collecting: %r %r %r\n' % (collector, name, obj)).encode('utf-8'))
+
+    return MypyDataSuite(name, parent=collector)
+
+
+class MypyDataSuite(pytest.Class):
+    def collect(self):
+        for case in self.obj.cases():
+            yield MypyDataCase(case.name, self, case)
+
+
+class MypyDataCase(pytest.Item):
+    def __init__(self, name: str, parent: MypyDataSuite, obj: DataDrivenTestCase):
+        self.skip = False
+        if name.endswith('-skip'):
+            self.skip = True
+            name = name[:-len('-skip')]
+
+        super().__init__(name, parent)
+        self.obj = obj
+
+    def runtest(self):
+        if self.skip:
+            pytest.skip()
+        update_data = self.config.getoption('--update-data', False)
+        self.parent.obj(update_data=update_data).run_case(self.obj)
+
+    def setup(self):
+        self.obj.set_up()
+
+    def teardown(self):
+        self.obj.tear_down()
+
+    def reportinfo(self):
+        return self.obj.file, self.obj.line, self.obj.name
+
+    def repr_failure(self, excinfo):
+        if excinfo.errisinstance(SystemExit):
+            # We assume that before doing exit() (which raises SystemExit) we've printed
+            # enough context about what happened so that a stack trace is not useful.
+            # In particular, uncaught exceptions during semantic analysis or type checking
+            # call exit() and they already print out a stack trace.
+            excrepr = excinfo.exconly()
+        else:
+            self.parent._prunetraceback(excinfo)
+            excrepr = excinfo.getrepr(style='short')
+
+        return "data: {}:{}\n{}".format(self.obj.file, self.obj.line, excrepr)
+
+
+class DataSuite:
+    @classmethod
+    def cases(cls) -> List[DataDrivenTestCase]:
+        return []
+
+    def run_case(self, testcase: DataDrivenTestCase) -> None:
+        raise NotImplementedError
