--- mypy/test/update.py
@@ -1,8 +1,0 @@
-import pytest
-
-
-def pytest_addoption(parser):
-    group = parser.getgroup('mypy')
-    group.addoption('--update-data', action='store_true', default=False,
-                    help='Update test data to reflect actual output'
-                         ' (supported only for certain tests)')
